refactor: move to controller
