--- packages/zwave-js/src/lib/driver/_Types.ts
@@ -1,11 +1,0 @@
-export interface FirmwareUpdateFileInfo {
-	target: number;
-	url: string;
-	integrity: `sha256:${string}`;
-}
-
-export type FirmwareUpdateInfo = {
-	version: string;
-	changelog: string;
-	files: FirmwareUpdateFileInfo[];
-};
